requirements.txt
streamlit
google-cloud-aiplatform
pillow
python-dotenv